refactor(gateway): move test stream observer class
Move the TestStreamObserver class to a more appropriate
test package.
