+++ gateway/src/test/java/io/camunda/zeebe/gateway/api/job/StreamActivatedJobsTest.java
@@ -10,7 +10,6 @@ package io.camunda.zeebe.gateway.api.job;
 import static org.assertj.core.api.Assertions.assertThat;
 
 import io.camunda.zeebe.gateway.api.util.GatewayTest;
-import io.camunda.zeebe.gateway.api.util.TestStreamObserver;
 import io.camunda.zeebe.gateway.protocol.GatewayOuterClass.ActivatedJob;
 import io.camunda.zeebe.gateway.protocol.GatewayOuterClass.StreamActivatedJobsRequest;
 import io.camunda.zeebe.protocol.impl.record.value.job.JobRecord;
